[kirkstone] Remove deprecated use of += with other overrides
As of Kirkstone, the use of += in combination with append,
prepend, and remove is no longer allowed. All uses of these
overrides will be folded into eachother, so there is no need
to specify the +=. As an example,

DEPENDS = "abc"
DEPENDS:append = " def"
DEPENDS:append = " ghi"

becomes "abc def ghi". Note the required space.