[kirkstone] ocaml: Upgrade to 4.05 and refresh
Upgrade from 4.04 to 4.05 to pull in fixes for gcc10 and
glib 2.34. Note that the 4.05 branch no longer has new tags,
but it still gets backports, so we're cloning from the head
of the 4.05 stable branch. Clone directly from github to use
the branch.

Also, add a few patches to suppress a few warnings and get
us building. Note that we should eventually update to a newer
ocaml release as this one is still quite old. This is just a
temporary measure to get us up and running in kirkstone.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>